Refactor Game component: extract game logic into separate utility functions for handling target hits, misses, health depletion, game over conditions, key events, score updates, and game restarts.
