+++ app/pages/asteroid/_comp/Game/handleMiss.js
@@ -1,0 +1,4 @@
+export const handleMiss = ({ setMisses, onMiss }) => {
+  setMisses((prevMisses) => prevMisses + 1);
+  if (onMiss) onMiss();
+};
