+++ app/pages/asteroid/_comp/Game/updateScore.js
@@ -1,0 +1,3 @@
+export const updateScore = ({ hits, misses, setScore }) => {
+  setScore(hits * 100 + Math.round((hits / (hits + misses || 1)) * 100));
+};
